webserver/requirements.txt
flask
mysql-connector-python
flask_limiter
requests
flask-cors
